add security, cinfig it , config jwt

+++ src/main/java/Moodle/Repositories/CoursesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface CoursesRepository extends JpaRepository<Courses,Long> {
+public interface CoursesRepository extends JpaRepository<Courses,Integer> {
 }
