+++ src/main/java/Moodle/Repositories/FilesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface FilesRepository extends JpaRepository<Files,Long> {
+public interface FilesRepository extends JpaRepository<Files,Integer> {
 }
